1st commit for milestone

+++ angular-jwt-master/src/app/pages/dashboard/upload-documents/upload-documents.component.ts
@@ -8,7 +8,10 @@ import { environment } from '@environments/environment.prod';
 import { LocalDataSource } from 'ng2-smart-table';
 import { BsModalRef } from 'ngx-bootstrap/modal';
 import { ToastrService } from 'ngx-toastr';
-import { Subscription } from 'rxjs';
+import { Subscription, Subject } from 'rxjs';
+import { MatDialog } from '@angular/material/dialog';
+import { ConfirmationAlertComponent } from '@app/shared/_components/confirmation-alert/confirmation-alert.component';
+import { takeUntil } from 'rxjs/internal/operators/takeUntil';
 declare var $: any;
 @Component({
   selector: 'app-upload-documents',
@@ -60,7 +63,7 @@ export class UploadDocumentsComponent implements OnInit, OnDestroy {
         valuePrepareFunction: (OCDate) => {
           var raw = new Date(OCDate);
           if (raw) {
-            return this.datePipe.transform(raw, 'dd/MM/yyyy ');
+            return this.datePipe.transform(raw, 'dd/MM/yyyy');
           }
         }
       },
@@ -79,24 +82,64 @@ export class UploadDocumentsComponent implements OnInit, OnDestroy {
       }
     },
   };
+  archiveSettings = {
+    actions: false,
+    columns: {
+      srNo: {
+        title: 'Sr No',
+        filter: false,
+        valuePrepareFunction: (cell, row) => { return cell }
+      },
+      documentname: {
+        title: 'Document Name',
+        filter: false,
+        type: 'custom',
+        renderComponent: CustomRendererFileComponent
+      },
+      notes: {
+        title: 'Notes',
+        filter: false
+      },
+      uploadedby: {
+        title: 'Uploaded By',
+        filter: false
+      },
+      uploadeddate: {
+        title: 'Uploaded Date',
+        filter: false,
+        valuePrepareFunction: (OCDate) => {
+          var raw = new Date(OCDate);
+          if (raw) {
+            return this.datePipe.transform(raw, 'dd/MM/yyyy');
+          }
+        }
+      },
+    },
+  };
+  unsubscribeAll: Subject<boolean> = new Subject<boolean>();
+  isClosed: boolean = false;
 
   constructor(private dashboardService: DashboardService, private router: Router, private datePipe: DatePipe,
     private authenticationService: AuthenticationService, private route: ActivatedRoute, private toasterService: ToastrService,
     public bsModalRef: BsModalRef
   ) {
-    this.currentUser$ = this.authenticationService.currentUserSubject.subscribe(data => {
+    this.authenticationService.currentUserSubject.pipe(takeUntil(this.unsubscribeAll)).subscribe(data => {
       if (data != null) {
         this.currentUser = data;
       }
     });
-    this.ocObj$ = this.dashboardService.selectedObj.subscribe(data => {
-      if (data) {
+    this.dashboardService.selectedObj.pipe(takeUntil(this.unsubscribeAll)).subscribe(data => {
+      if (data != null) {
         this.ocObj = data;
+        console.log('ocobj', this.ocObj)
         if (this.ocObj.OCDate && this.ocObj.OCDate.formatted) {
           this.ocObj.OCDate = this.ocObj.OCDate.formatted;
           this.isModal = true;
         } else {
-          this.ocObj.OCDate = this.datePipe.transform(this.ocObj.OCDate, 'yyyy-MM-dd');
+          this.ocObj.OCDate = this.datePipe.transform(new Date(this.ocObj.OCDate), 'dd/MM/yyyy');
+          if (this.ocObj.Status.name == 'Closed') {
+            this.isClosed = true;
+          }
         }
       }
     })
@@ -111,7 +154,8 @@ export class UploadDocumentsComponent implements OnInit, OnDestroy {
   }
 
   ngOnDestroy() {
-    this.currentUser$.unsubscribe();
+    this.unsubscribeAll.next(true);
+    this.unsubscribeAll.complete();
   }
   getDocuments() {
     const body = {
@@ -167,7 +211,7 @@ export class UploadDocumentsComponent implements OnInit, OnDestroy {
     }
   }
   onBack() {
-    this.router.navigate(['/pages/dashboard']);
+    this.router.navigate(['/pages/oc-list']);
   }
 }
 @Component({
@@ -191,27 +235,39 @@ export class CustomRendererFileComponent implements OnInit {
 @Component({
   selector: 'app-custom-renderer-file-delete',
   template: `<span (click)="onDelete()" class="font-medium-1 mr-2" style="cursor:pointer;color:blue;font-size:16px" data-toggle="tooltip" data-placement="top" title="Delete"><i class="fa fa-trash-o fa-lg" aria-hidden="true"></i></span>`
-  
+
 })
 export class CustomRendererFileDeleteComponent implements OnInit {
 
-  constructor(private dashboardService: DashboardService, private toasterService: ToastrService, private router: Router) {
+  constructor(private dashboardService: DashboardService, public dialog: MatDialog, private toasterService: ToastrService, private router: Router) {
   }
   @Output() save: EventEmitter<any> = new EventEmitter();
   renderValue: string;
   @Input() value: string | number;
   @Input() rowData: any;
   ngOnInit() {
-    
+
   }
   onDelete() {
-    this.dashboardService.deleteDocument(this.value).subscribe(res => {
-      if (res.status === 'success') {
-        this.toasterService.success(res.message);
-        this.save.emit(true);
-      } else {
-        this.toasterService.error(res.message);
+    const dialogRef = this.dialog.open(ConfirmationAlertComponent, {
+      panelClass: 'app-dialog',
+      width: '300px',
+    });
+
+    dialogRef.componentInstance.title = `Delete document`;
+    dialogRef.componentInstance.message = `Are you sure, you want to delete this document?`;
+
+    dialogRef.afterClosed().subscribe(result => {
+      if (result) {
+        this.dashboardService.deleteDocument(this.value).subscribe(res => {
+          if (res.status === 'success') {
+            this.toasterService.success(res.message);
+            this.save.emit(true);
+          } else {
+            this.toasterService.error(res.message);
+          }
+        })
       }
-    })
+    });
   }
 }
